% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{pre-process World Bank data}
\usage{
clean_data(data, value_name)
}
\arguments{
\item{data}{data with wide frame}

\item{value_name}{change name of value stored}
}
\description{
a function that turns data frame from wide to long format and contain variables - iso, year, data
}
\examples{
clean_data(maternal_mortality, "maternal_mortality")
}
\author{
Soohyun Yoon
}
